refactor(docs): rename Renewal Interaction to Default Interaction
