+++ apps/insight-viewer-docs/containers/Interaction/Tabs.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { Tabs, TabList, TabPanels, Tab, TabPanel } from '@chakra-ui/react'
 import Base from './Base'
 import Custom from './Custom'
-import Renewal from './Renewal'
+import DefaultInteraction from './DefaultInteraction'
 
 export default function App(): JSX.Element {
   const [active, setActive] = useState(0)
@@ -16,7 +16,7 @@ export default function App(): JSX.Element {
       <TabList>
         <Tab>Base Interaction</Tab>
         <Tab className="custom-tab">Custom Interaction</Tab>
-        <Tab className="renewal-tab">Renewal Interaction</Tab>
+        <Tab className="default-tab">Default Interaction</Tab>
       </TabList>
 
       <TabPanels>
@@ -27,7 +27,7 @@ export default function App(): JSX.Element {
           {active === 1 && <Custom />}
         </TabPanel>
         <TabPanel p={0} pt={6}>
-          {active === 2 && <Renewal />}
+          {active === 2 && <DefaultInteraction />}
         </TabPanel>
       </TabPanels>
     </Tabs>
